Inbound close handler on RequestBody
