+++ Sources/HummingbirdCore/Request/Request.swift
@@ -34,9 +34,6 @@ public struct Request: Sendable {
     @inlinable
     public var headers: HTTPFields { self.head.headerFields }
 
-    @usableFromInline
-    let iterationState: RequestIterationState?
-
     // MARK: Initialization
 
     /// Create new Request
@@ -50,7 +47,6 @@ public struct Request: Sendable {
         self.uri = .init(head.path ?? "")
         self.head = head
         self.body = body
-        self.iterationState = nil
     }
 
     /// Create new Request
@@ -59,13 +55,11 @@ public struct Request: Sendable {
     ///   - bodyIterator: HTTP request part stream
     package init(
         head: HTTPRequest,
-        bodyIterator: NIOAsyncChannelInboundStream<HTTPRequestPart>.AsyncIterator,
-        supportCancelOnInboundClosure: Bool
+        bodyIterator: NIOAsyncChannelInboundStream<HTTPRequestPart>.AsyncIterator
     ) {
         self.uri = .init(head.path ?? "")
         self.head = head
         self.body = .init(nioAsyncChannelInbound: .init(iterator: bodyIterator))
-        self.iterationState = supportCancelOnInboundClosure ? .init() : nil
     }
 
     /// Collapse body into one ByteBuffer.
@@ -80,11 +74,6 @@ public struct Request: Sendable {
         let byteBuffer = try await self.body.collect(upTo: maxSize)
         self.body = .init(buffer: byteBuffer)
         return byteBuffer
-    }
-
-    @inlinable
-    package func getState() async -> RequestIterationState.State? {
-        await self.iterationState?.state
     }
 }
